Extract SettleRequest
